vfs: stop reporting an I/O error as a four-billion-byte successful read
VFS-28.  Filesystem backends report failure by returning a negated errno
cast through size_t -- (size_t)-EIO is 0xFFFFFFFA.  read_fs()/write_fs()
returned that verbatim as an unsigned byte count, so an I/O error reached
the caller as a ~4 GiB SUCCESSFUL TRANSFER.  Callers that compared against
an expected length happened to catch it; callers that used the value as a
length did not.

Both now return ssize_t.  Rather than re-type every backend's read/write
hook -- dozens of functions across ext2, fat, exfat, minix, sysv, udf,
procfs, devfs, sysfs, shmfs, 9p and the device nodes -- the error range is
recognised on the way out, the way Linux's IS_ERR_VALUE does: a return of
>= (size_t)-4095 is an errno, not a count.  A transfer can never
legitimately be that large, since `size` is bounded far below 4 GiB - 4096,
so the range is unambiguous and no backend has to change.

Two related bugs in the same functions:

  - atime was stamped before the result was examined, so a read that FAILED
    still advanced the access time.  Now only a successful read does, and
    likewise mtime/ctime for writes.

  - a node with no ->read returned 0, indistinguishable from a clean EOF, so
    a caller looping until 0 treated a directory or a write-only device as
    an empty file.  That is -EINVAL now, as is a NULL node.

-Wsign-compare found the three call sites that compared the result against
an unsigned length -- exactly the places a negative return would have been
misread -- in the ELF header, phdr and dynamic-section probes.  All three
now cast the bound, so a short read and an error both fail the test where
an error previously compared as huge and passed it.  The remaining ~67 call
sites test `!= EXPECTED_SIZE`, which behaves identically under a signed
return.

Verified: builds clean tree-wide with -Werror, and boots to rc.d/60-sdm.
Every ext2 superblock, inode, directory block and ELF image loaded during
boot passes through read_fs, so the new path is thoroughly exercised.

+++ sys/kern/syscall.c
@@ -182,7 +182,10 @@ static int elf_dt_needs_substrate_libc(fs_node_t *node, const Elf32_Ehdr *eh) {
     if (eh->e_phentsize != sizeof(Elf32_Phdr) || phnum == 0 || phnum > 32)
         return 0;
     phbytes = (size_t)phnum * sizeof(Elf32_Phdr);
-    if (read_fs(node, eh->e_phoff, phbytes, (uint8_t *)ph) < phbytes)
+    /* [VFS-28] read_fs is signed now: a short read AND an error both fail
+     * this test, where an error used to compare as a huge unsigned value and
+     * pass it. */
+    if (read_fs(node, eh->e_phoff, phbytes, (uint8_t *)ph) < (ssize_t)phbytes)
         return 0;
 
     for (uint16_t i = 0; i < phnum; i++) {
@@ -195,7 +198,7 @@ static int elf_dt_needs_substrate_libc(fs_node_t *node, const Elf32_Ehdr *eh) {
         return 0;
     if (dyn_sz > sizeof(dyn))
         dyn_sz = sizeof(dyn);
-    if (read_fs(node, dyn_off, dyn_sz, dyn) < dyn_sz)
+    if (read_fs(node, dyn_off, dyn_sz, dyn) < (ssize_t)dyn_sz)
         return 0;
 
     /* First pass: locate DT_STRTAB (a virtual address) and map it to a file
@@ -241,7 +244,7 @@ static int vfs_node_is_substrate_object(fs_node_t *node) {
     const Elf32_Ehdr *eh = (const Elf32_Ehdr *)hdr;
     if (!node || (node->flags & 0x7) != FS_FILE)
         return 0;
-    if (read_fs(node, 0, sizeof(hdr), hdr) < sizeof(hdr))
+    if (read_fs(node, 0, sizeof(hdr), hdr) < (ssize_t)sizeof(hdr))
         return 0;
     if (hdr[0] != 0x7f || hdr[1] != 'E' || hdr[2] != 'L' || hdr[3] != 'F')
         return 0;
